Added quickex proof of concept
The idea here is to make regex determination verbal using a recursive
descent parser. I'm not sure if it's wholly possible once I get to
negative lookbehind assertions and stuff, but hopefully it will be!

Ideally the project will be easily deployable locally or on an intranet
to help developers decode complex regex that they try to create or find
in a source code file that is not properly commented.